+++ module/data/item-condition.mjs
@@ -1,0 +1,5 @@
+import ProwlersParagonsItemBase from "./base-item.mjs";
+
+// This is really just a container for active effects
+export default class ProwlersParagonsCondition extends ProwlersParagonsItemBase {
+}