+++ playwright.config.js
@@ -1,5 +1,6 @@
 // @ts-check
 import { defineConfig, devices } from '@playwright/test';
+import path from 'path';
 
 /**
  * Read environment variables from file.
@@ -8,6 +9,8 @@ import { defineConfig, devices } from '@playwright/test';
 // import dotenv from 'dotenv';
 // import path from 'path';
 // dotenv.config({ path: path.resolve(__dirname, '.env') });
+
+const authFile = path.join(__dirname, 'playwright/.auth/user.json');
 
 /**
  * @see https://playwright.dev/docs/test-configuration
@@ -21,10 +24,10 @@ export default defineConfig({
   /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
   /* Opt out of parallel tests on CI. */
-  workers: process.env.CI ? 1 : undefined,
+  workers: 4,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: 'list',
-  timeout: 30000,
+  timeout: 20000,
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
     /* Base URL to use in actions like `await page.goto('/')`. */
@@ -37,13 +40,19 @@ export default defineConfig({
   /* Configure projects for major browsers */
   projects: [
     {
+      name: 'setup',
+      use: { ...devices['Desktop Chrome'], channel: 'chromium' },
+      testMatch: /.*\.setup\.js/,
+    },
+    {
       name: 'cleanup',
       use: { ...devices['Desktop Chrome'], channel: 'chromium' },
       testMatch: /global\.teardown\.js/,
     },
     {
       name: 'chromium',
-      use: { ...devices['Desktop Chrome'], channel: 'chromium' },
+      use: { ...devices['Desktop Chrome'], channel: 'chromium', storageState: authFile },
+      dependencies: ['setup'],
       teardown: 'cleanup',
     },
     // {