+++ tests/global.teardown.js
@@ -1,7 +1,7 @@
 import { test as teardown } from '@playwright/test';
 import { login } from './login-helper';
 
-teardown('delete actors and tokens', async ({ page }) => {
+teardown('delete actors, tokens, and items', async ({ page }) => {
     await login(page);
     await page.getByRole('tab', { name: 'Actors' }).click();
     await page.evaluate(async () => {
@@ -16,6 +16,10 @@ teardown('delete actors and tokens', async ({ page }) => {
         const scene = game.scenes.getName('Foundry Virtual Tabletop');
         for (const token of scene.tokens) {
             await token.delete();
+        }
+
+        for (const item of game.items) {
+            await item.delete();
         }
 
     });